ai-business-analyst/Backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
pydantic==2.5.3
pypdf==4.2.0
reportlab==4.2.2
requests==2.32.3
